requirements.txt
langchain-chroma
langchain-openai
python-dotenv
pyyaml
chromadb
rich
